+++ changelog.md
@@ -5,8 +5,8 @@ Date: 04 April 2026
 
 - **Feature**:
     - Added hover effect on partner logos.
-    - Added /services and /legals pages.
     - Added /about/j42l and /about/members pages. Url /about will redirect to /about/j42l.
+    - Added multi language
 
 - **Refactor**:
     - Semantic HTML